requirements.txt
streamlit==1.37.0

google-generativeai==0.8.5

langchain==0.3.7
langchain-core==0.3.45
langchain-google-genai==2.0.10
langchain-chroma==0.1.4

langgraph==0.2.34

python-dotenv==1.0.1

pydantic==2.9.2
tenacity==8.5.0

PyGithub==2.9.0
pymongo==4.16.0

plotly==6.6.0